Add utility methods to retrieve paths of files/folders within bundles
As GDA-9 server can be run either from eclipse or an exported product
(and so can the client for that matter), the paths of, say, the scripts
folder within uk.ac.gda.core change based on the active launch method.
These methods wrap the eclipse FileUtils.find method which can retrieve
the correct paths for files and folders from the currently active bundle
context. 

This allows various hard coded paths in the code (which would by default
be incorrect in the exported product case) to be replaced with dynamic
determination of the correct paths at runtime based on the bundle name
and relative path to the file/folder within it.

This in turn will allow much of the dls-root folder (which was set up as
a quick and dirty way round this problem) to be removed from the
exported server product.

Change-Id: I6857a63601511618b33ce56b9a9d6da4f8b92707