Add Windows path handling to File resolution and re-factor
On Windows boxes, calling getPath on a URL returned by the FileLocator
class methods yields a windows style path preceded by a forward slash
and using forward slashes throughout e.g. for the path C:\windows it
would return /C:/windows. The leading slash needs to be stripped off to
allow the rest of the file resolution methods to work consistently with
both OSs.

Also the body of resolveBundleFile is factored out into a mathod so that
it can be called from other classes.

Change-Id: Ibc9e4f5f5abdaa3275531066cf2957e9e2a52153