DASCTEST-101 GDA-5573 Remove com.springsource.org.apache.xerces jar from uk.ac.gda.common
This jar has been replaced by a dependency on the org.apache.xerces
plugin already in the tp.

It questionably if we even need this, the only class using it is
XMLObjectConfigFileValidator which seems to be aimed at validating
Castor files which we don't use any more. However there are a few uses
of it in the workspace so I didn't remove it here.

Change-Id: Ibed1f90d253150ea2d38293e7d1d543270dee71d