Don't lose the GDA_SOURCE property when a Logback LoggerContext is reset
The GDA_SOURCE property was only added to a LoggerContext once, when the GDA
process starts up.

If the LoggerContext was subsequently reset - for example, if the configuration
file(s) were changed, and were automatically reloaded - this property would be
lost.

Initially log messages would be OK:

  2017-03-21 13:53:55,967 INFO  [objectserver/3623@ws341.diamond.ac.uk] sample_control - HI

But once the context had been reset, the process name would be missing:

  2017-03-21 13:54:01,102 INFO  [null/3623@ws341.diamond.ac.uk] sample_control - HI

Add a LoggerContextListener that restores the property when the context is reset.

Change-Id: I86c5d166bb1c94189ad34ae5682dcc3f5dab6478