DAQ-470 Make Local Property loading explicit
Currently there is no explicit load method for LocalProperties, rather
they all get loaded on the first call to any method of the class via the
static initialiser. While this is fine, it doesn't really convey what's
actually happening so I've added an explicit (blank) load method to
LocalPropertiesfor this purpose.

This only exists to trigger the static initialiser in a more
meaningful way whilst retaining the Thread safety it guarantees.

Also make the common config method in LogbackUtils require the logging
config filename to be passed in to remove the need to get it from
LocalProperties. The server and client methods are then rearranged so
that only the client one uses LocalProperties.

Change-Id: Ica7a3eb53d6580e671e71f3b867938337fa38189