Fix 2 warnings for outdated JRE in .classpath
classpath specifies a JRE version older than the workspace default
(current version is 1.8). Changing the .classpath file to take the JRE
default version from the workspace instead of an explicit definition
fixes the warning, and should prevent it happening again when the
workspace gets updated to a newer JRE version.

Change-Id: I10ac48517bd3f346ef5d302e2544ed790230b4ae